-Cumulative squares finally implemented.
-Now squares do gets cumulative. Later in the software development i'll implement the option for the user to choice how many measurements can take place in a single location, so the count variable stands for this. Also I've added stuff to the databasemanager for the refreshment of the values.

+++ app/src/main/java/com/example/lam_project/LTESignalPainter.java
@@ -27,7 +27,7 @@ public class LTESignalPainter {
 
     // Method to paint the square based on the LTE signal strength
     public static void paintSquareByLTESignalStrength(MapView map, Polygon square,
-                                                      int signalStrength, int mode,
+                                                      double signalStrength, int mode,
                                                       double squareSizeMeters) {
         if (map == null || square == null)
             return;
